app/requirements.txt
Django==4.0.1
psycopg2-binary==2.9.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
aiogram==2.19
drf-yasg==1.20.0
celery==5.2.3
redis==4.1.2
boto3==1.20.49
fleep==1.0.1
numpy==1.22.3
pandas==1.4.1
sklearn
xgboost==1.5.2
twilio==7.7.1
